cleanup code again
