+++ core/views.py
@@ -1,23 +1,16 @@
 from django.shortcuts import redirect, render
-
+from core.lib import create_booking
 from core.models import Booking, Room
 
 def book(request):
     if request.method == 'POST':
         create_booking(request)
         return redirect('index')
-
-    rooms = Room.objects.all()
-    return render(request, 'book.html', {'rooms': rooms})
+    else: 
+        return render(request, 'book.html', {'rooms': Room.objects.all()})
 
 
 def index(request):
-    bookings = Booking.objects.all()
-    
-    return render(request, 'index.html', {'bookings': bookings})
+    return render(request, 'index.html', {'bookings': Booking.objects.all()})
 
-def create_booking(request):
-    organiser, date, start, end, title, details, room_id = request.POST.get('organiser'), request.POST.get('date'), request.POST.get('start'), request.POST.get('end'), request.POST.get('title'), request.POST.get('details'), request.POST.get('room')
-    room = Room.objects.get(id=room_id)
-    booking = Booking.objects.create(organiser=organiser, date=date, start_time=start, end_time=end, room=room,  title=title, details=details)
-    booking.save()
+
